+++ common/hydracore/call_server.go
@@ -7,8 +7,8 @@ const (
 	callClientEnabled = false
 	callServerEnabled = true
 	distributionRole  = "vps"
-	callWireMin       = 7
-	callWireMax       = 7
+	callWireMin       = 8
+	callWireMax       = 8
 )
 
 var (